move param check to QRNavigator.tsx
