+++ src/qrcode/QRCode.tsx
@@ -12,39 +12,11 @@ import colors from 'src/styles/colors'
 import fontStyles from 'src/styles/fonts'
 import variables from 'src/styles/variables'
 import { currentAccountSelector } from 'src/web3/selectors'
-import { Statsig } from 'statsig-react-native'
-import { QRCodeAppearance, QRCodeDataType, StatsigLayers } from 'src/statsig/types'
-import { LayerParams } from 'src/statsig/constants'
-import Logger from 'src/utils/Logger'
+import { QRCodeDataType } from 'src/statsig/types'
 
 interface Props {
   qrSvgRef: React.MutableRefObject<SVG>
   dataType: QRCodeDataType
-}
-
-const TAG = 'QRCode'
-
-function getExperimentParams(): {
-  qrCodeAppearance: QRCodeAppearance
-  qrCodeData: QRCodeDataType
-} {
-  const layerName = StatsigLayers.SEND_RECEIVE_QR_CODE
-  const { paramName: appearanceParamName, defaultValue: appearanceDefaultValue } =
-    LayerParams[layerName].qrCodeAppearance
-  const { paramName: dataParamName, defaultValue: dataDefaultValue } =
-    LayerParams[layerName].qrCodeData
-  try {
-    const statsigLayer = Statsig.getLayer(layerName)
-    const qrCodeAppearance = statsigLayer.get(appearanceParamName, appearanceDefaultValue)
-    const qrCodeData = statsigLayer.get(dataParamName, dataDefaultValue)
-    return { qrCodeAppearance, qrCodeData }
-  } catch (error) {
-    Logger.warn(TAG, 'error getting Statsig experiment', error)
-    return {
-      qrCodeAppearance: appearanceDefaultValue,
-      qrCodeData: dataDefaultValue,
-    }
-  }
 }
 
 export const mapStateToProps = (state: RootState) => ({
@@ -56,7 +28,6 @@ export const mapStateToProps = (state: RootState) => ({
 export default function QRCodeDisplay({ qrSvgRef, dataType }: Props) {
   const data = useSelector(mapStateToProps, shallowEqual)
   const qrContent = useQRContent(dataType, data)
-  Logger.debug(TAG, `experiment params: ${JSON.stringify(getExperimentParams())}`)
   return (
     <SafeAreaView style={styles.container}>
       <AvatarSelf iconSize={64} displayNameStyle={fontStyles.h2} />
